Move settings to settings route
