+++ admin/package.json
@@ -32,6 +32,7 @@
     "@types/node": "^14.14.6",
     "@types/react": "^16.9.55",
     "@types/react-dom": "^16.9.9",
+    "@types/react-router-dom": "^5.1.6",
     "axios": "^0.20.0",
     "express": "^4.17.1",
     "prop-types": "^15.7.2",
